refactor: single integration test file for the gateway
