+++ gateway/ln-gateway/Cargo.toml
@@ -24,10 +24,6 @@ path = "src/bin/cln_extension.rs"
 name = "gatewayd-integration-tests"
 path = "tests/integration_tests.rs"
 
-[[test]]
-name = "gateway-client-tests"
-path = "src/ng/tests.rs"
-
 [dependencies]
 anyhow = "1.0.66"
 async-stream = "0.3.5"
